requirements.txt
openai==1.88.0
PyYAML==6.0.2
PyYAML==6.0.2
qwen_vl_utils==0.0.11
torch==2.4.0
tqdm==4.67.1
transformers==4.50.3
transformers==4.49.0
